requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.34.0,<1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.5.2,<3.0.0
anyio>=4.9.0,<5.0.0
starlette>=0.46.2,<1.0.0
httpx[http2]>=0.28.1,<1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.9.0.post0,<3.0.0
structlog>=23.2.0
